Revert "use shared project to get rid of extra dll."
This reverts commit 8da5bdf5ac6a82838317cc07631706adadff7a69.
