--- NetworkSkins.Helpers/NetworkSkins.Helpers.projitems
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>0cd23759-b535-4c44-b174-a5cd863c5b76</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>NetworkSkins.Helpers</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)DelegateUtil.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)INSImplementation.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NSHelpers.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)NSImplementationWrapper.cs" />
-  </ItemGroup>
-</Project>
