+++ NetworkSkins.Intreface/INSImplementation.cs
